Bump Version 1.3.128

+++ source/ddb/file_io/locking.py
@@ -119,6 +119,9 @@ class lock:
     @staticmethod
     def check_pid(pid):        
         """ Check For the existence of a unix pid. """
+        if os.getpid()==pid:
+            lock.info("Lock","This Thread {0}".format(pid))
+            return True
         try:
             os.kill(pid, 0)
         except OSError:
